Frameworks: Gesture Lockscreen (1/2)
This adds the ability to us a hand drawn gesture as
a means to unlock the device.

In the case that the gesture is forgotton or not
being recognized correctly, the fallback will be the
same as when using patterns.

Androidx:
Some changes needed for 5.0

pimpmaneaton:
Some changes needed for 5.1

Change-Id: Iedaed283bfc317d4b14f21bad6fca1a3ffd4a190

Conflicts:
	core/java/com/android/internal/widget/LockPatternUtils.java