Remove carrier text from the keyguard(lockscreen)
MAY come back to this and add a separate switch but if not I like this better
If I don't have something custom set or even turned on at all then I don't wanna see
this everytime I unlock my phone. Unless it's set to something custom/funny I don't see the
point in the feature to begin with. Was request but now I'm cleaning up

Change-Id: Id56a938e0dc8948cb1f82306f00805594e8120ac

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/KeyguardStatusBarView.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java