Lock screen shortcuts color cleanup 2/2
Additions:
- Icons are using the lock screen icon color, (custom targets optional)
- Removed the saturation color filter for custom target icons
- Option to colorize all icons:
  - Disabled:
    - Custom target icons are shown without any modifications
    - Custom target touch feedbacks are using the pallete color of the icon, (no change)
  - Enabled:
  - Custom target icons and touch feedback are using the lock screen icon color

Pulled out of this commit, it was full locksreen shortcuts add and has different strings
so adapted to cm/cyanide setup without having to do a huge rework. Already test and working

Thanks to XXMrHyde, hard to separate sometimes but totally worth it, this guy is awesome.
I would CP but the differences in source make it much easier this way. Credits shown in the way he requested
XXMrHyde/android_frameworks_base@3194062

Change-Id: I124fb526d66f0114518c38a458e2d76d53ee3046