Navbar: Unpin App
This is probably the dirtiest hack/fix I've ever seen. Kinda embarased by it. But since the new Navbar customization commits, we've lost the ability to unpin an app since the system doesn't recognize holding the Back & Recents buttons together. So like the idiot I am, I created an action that the user can assign to a Navbar button. It's not a fix, but a hax.

If anyone has a better fix, I'm all ears.

PS2:
Create a different toast and card message for the use telling them how to unpin.
The next step is to find a way to assign the ACTION_UNPIN to a button by default.