LEDs Brightness [2/2]: Lights notifications brightness support
Implement the support of an overall brightness control for the LEDs.

LEDs Brightness and Custom values are separated
in an advanced section for a cleaner overview.

The setting is deactivated by default
and will be ignored by the unimplemented phones.
Current LibLights will simply not use the new variable.

Changes includes :
  frameworks/base
  packages/Apps/Settings

Device LibLights library rewritten for the SONY Huashan :
  https://github.com/AdrianDC/android_device_sony_huashan/tree/cm-12.1/libl$

Change-Id: I1c0de01b1c6a2a2cf1432028a2e69f90b2373b2c
Signed-off-by: AdrianDC <radian.dc@gmail.com>

Conflicts:
	core/java/android/provider/Settings.java
	services/core/java/com/android/server/notification/NotificationManagerService.java