GlobalMenu: Fix sound panel alignment for custom dpi modes.
When we use a lower dpi mode then the device its default dpi,
the power menu sound panel was aligned to the left.
This causes an issue because the ripple effects where centered
and the buttons where not centered in its view.
Thats looks terrible when the ripple effect is shown when clicked.

Signed-off-by: BlackDragon <blackdragon.fusionteam@gmail.com>