Status bar: Greeting, (1/2):
 Contributors:
rogersb11 and  BlackDragon,
(Original author: AlmightyMegadeth00)

Included:
- Show/Hide greeting
- Customizeable greeting text
- Timeout
- Text color

Additional changes:
- Use the complete width of the status bar for the greeting
- Center the greeting horizontal
- cange the default greeting from "Welcome to <Rom name>" to:
  "Welcome: <Rom name>, <Android version> (<Build id>), <Build date>",  e.g.:
  "Welcome: CyanideL, 5.1.1 (LMY48P), 2015-09-20"
- Hide "all" status bar content if greeting is shown
- Change timeout behavior,
  (timeout is the time the greeting is visible in full opaque now)
- Extent the Show/hide option by a possibility to only show the greeting
  once after the boot

Cyanide: Adapt to my clock setup and it is different for sure, change lyout to location and add Clock. to beginning of chosen style

The status bar battery bar needs some overwork to show the expected behavior,
(hidden when the greeting will be shown).
While fixing the battery bar for the greeting, I also noticed some more issues to fix
for the battery bar, so I decided to not touch the battery bar for now,
this should be fixed in a seperate commit.
(Will fix the bar with one of the next commits).